steptwo/requirements.txt
Django==1.9
gunicorn==19.6
pillow==2.1.0